Added unit test for addVertex.
Implemented testAddVertex() unit test. Also created two parameters for the unit
tests, "coordLoc" and "servers", which pass the coordinator locator and number
of master servers to the unit tests with -DcoordLoc="..." and -Dservers=N.

This unit tests is incomplete, since RAMCloudVertex currently has no way to
read its own properties in order to verify that the properties were written
correctly to RAMCloud.