Added MeasurementClient to run perf. experiments.
MeasurementClient is a main class that runs various micro benchmarks on the
graph database. Currently only the following benchmarks are implemented:

testAddVertexThroughput
testAddVertexLatency

testAddVertexThroughput is multithreaded to simulate multiple clients, and
testAddVertexLatency is singlethreaded to measure the latency distribution of
one client accessing the database (in other words, the latency distribution of
addVertex on an unloaded database).

testAddVertexLatency generates a reverse CDF plot using JFreeChart.