Made MeasurementClient executable by cluster.py
Added standard RAMCloud client program arguments to MeasurementClient and also
wrote a launch script for the MeasurementClient to be handed to cluster.py as
the client argument. This launch script is needed because stock cluster.py
inserts RAMCloud client arguments at the beginning of the client argument list,
which for a java program is the classpath and main class to execute, which are
required to be the first arguments. The launch script simply reverses this
order.

The goal in doing this is to leverage cluster.py's ability to launch clients on
multiple servers so as to generate distributed workloads.