Updated MeasurementClient to be distributed.
MeasurementClient is now able to execute distributed tests using the same
master / slave style of communication and control found in the RAMCloud cluster
performance benchmarks. The client with clientIndex=0 is the master, while the
rest are the slaves. The master issues commands to slaves and receives
responses back from them through a signaling table in RAMCloud containing
various registers for storing the command, its arguments, slaves return status,
and slaves return values.

Currently there seems to be an occasional issue with timeouts firing during
tests inolving around 7 or more clients, but that may be an issue with some of
the hardcoded timeout constants being too tight.