Added vertices method implementation.
Added the vertices() method implementation for RAMCloudGraph. The
implementation of this method reads what is necessary from RAMCloud to
construct RAMCloudVertex objects that contain the vertex's id and label.
Other information, such as incident edges or properties is not cached in the
returned vertex object. Thus, any calls to Vertex.edges or Vertex.properties
will result in a remote call to RAMCloud.

The caching strategy (or non-caching) strategy employed by Vertex objects
carries with it certain tradeoffs. Caching inherently introduces the
possibility of reading stale data, although the probability of that being true
is a function of how much time has passed since the last read and how
frequently the data is changing. A non-caching strategy means that reads always
reflect our own updates to the database, as well as updates we have been
informed about by others before we began reading, but may result in
unnecessary round trips to the database when the data has not changed.

There is also the question of when to cache in data, and how much. Prefetching
data from the cluster may be useful to save round trips later, but is wasted
if the data is not used.

Which tradeoffs to make depends almost entirely on how the database is
accessed, and so remains to be seen. In the meantime, our first approach is to
employ a strictly on-demand approach, where only the data that is needed to
satisfy the current query is fetching from RAMCloud. Exploring the spectrum of
possibilities that lies between this and a full caching scheme remains for
future work.