Implemented property and properties for vertices.
Found it simpler to write all methods that read/write RAMCloud values and
serialize / deserialize them in the RAMCloudGraph class. This centralizes the
logic for these operations, and avoids the need to have a copy of the RAMCloud
object in every Vertex and Edge object.

If a better way of organizing this logic becomes apparent then the code can be
refactored.