Attached RunListener to unit tests to see progress
The StructureStandardTest Suite includes many tests and takes a long time to
finish, making it useful to see intermediate progress. The
UnitTestProgressReporter RunListener does exactly this so we can see where
along the test path we are.

Also modified the pom file to set the GREMLIN_TESTS environment variable to
select for running only the GraphTest set of tests for now.