Fixed bugs revealed by GrahTest unit tests.
RAMCloudGraph now passes all but one test in the GraphTest set of unit tests.
The remaining tests checks for a specific exception to be thrown in the edges
method which is currently not supported, and the unit test doesn't check this
before running said test.

Also added a RAMCloudHelper, and put a couple methods in it to help us
serialize and deserialize the RAMCloud object blobs that contain an element's
set of properties.