Modified testAddVertexLatency to be distributed.
testAddVertexLatency can now be run with multiple clients providing load to the
system. The master client waits for slaves to be running load, then takes
latency samples one at a time, and records the results in a file, one file per
run to be used as data input for gnuplot.