Changed vertex ID format.
Previously IDs were generated by atomically incrementing a counter in a
RAMCloud table. This will eventually become a bottleneck as all clients must
increment this counter in order to create new elements in the graph.

Now IDs are split into two parts: a client ID part, and a client local ID
counter part. When a new client is fired up, it grabs a new unique ID using a
RAMCloud counter and uses that ID for its lifetime. When this client wants to
create a new element in RAMCloud, it generates a globally unique ID by
concatenating its client id with a local counter.

For the time being, the generated ID is a string of two hexadecimal numbers
with a hyphen in between. In the future, this will be changed to a more
efficient encoding, but this works for the time being.