Changed internal vertex id to be of type byte[].
Previously, RAMCloud{Vertex, Edge, Element}'s id field was of type String. But
this turned out to be inconvenient for the implementation of edge lists, since
the edge list format, for efficiency, will use a 16B byte array containing two
longs to identify the {in,out} vertex. Instead of parsing strings to get the id
into this format, we store the id in this format natively.

Provisionally, we convert this byte array to a string representation when
constructing RAMCloud object keys for accessing different parts of a vertex.
This is storage inefficient, and will be changed to direclty use the id byte
array in the future as a part of the RAMCloud object key. For now the string
representation helps with debugging.