Implemented FEATURE_ADD_EDGES.
Implemented add edge functionality. Delete edges (and vertices for that matter)
are still unimplemented. Currently passes all unit tests in GraphTest.

Notes:
 - Vertices currently have their labels stored among their properties, which
means that to get just a vertex's label one must read all of their properties.
An example is, given an edge, finding the in and out vertices. To return the
vertices only the label needs to be fetched for each vertex, not all the
properties. Fetching a vertex given a specific ID is another example (fetching
the properties is not strictly needed, only the label). And lastly, filtering
for neighbors of a given type would be made more efficient by leaving vertex
labels to be individualy fetchable.
 - Fetching all vertices in the graph involves an enumeration of the vertex
table, which contains several ramcloud objects per vertex. This results in
many reads per vertex in the table.