Removed unnecessary data deletion in clear method.
Previously, the clear() method in RAMCloudGraphProvider would create a new
RAMCloudGraph object if the provided graph argument was null, and then proceed
to delete all the data in RAMCloud. However, given the way that a GraphProvider
is used by the test suite, that's not actually necessary because the same clear
method will always be called at the end of a test.