Moved vertex label to its own ramcloud object.
Previously, labels were stored with the vertex's properties, but there were
many cases where only the label was needed to satisfy the query. Placing the
label in a separate object reduces the amount of data that has to be read to
satisfy these queries, and increases the objects that need to be created at
vertex creation time by 1.