Removed heavy weight initialization from constructor.
Gremlin-Test unit tests must open a new graph before checking the feature
requirements for each test, which for RAMCloudGraph involved making a remote
connection to the RAMCloud cluster, creating several tables, and grabbing a
unique client ID. In the event that the graph does not support the required
features for the unit test, the graph is then closed and all that work is
wasted, making unit tests take much longer than needed to run, since none of
the work involved in creating a new RAMCloudGraph is needed to return the
features.

To fix this, the heavy weight initialization parts of RAMCloudGraph object
construction have been moved to a private initialize() method, which is
triggered to run once at the first invocation of any method that needs to
interact in any way with RAMCloud.

This has made the unit tests run much, much faster :)