Made Element.id() play nice with unit tests.
Previously, certain unit tests were not passing because they were using the
string representation of Element.id() and of graphProvider.convertId to
determine whether or not a read element has the correct id, but this assumes
that what is returned by Element.id() has a string representation in the first
place. Since all RAMCloud elements were using a byte[] as their ID, comparing
ids based on their string representations did not work. This commit changes the
return value of RAMCloudVertex.id() to be a BigInteger, which does have a
string representation. RAMCloudGraphProvider.convertId() also returns a
BigInteger, and now the addVertex() and vertices() methods of RAMCloudGraph
can take a BigInteger as the supplied Id or creating or looking up a vertex.

This weirdness is essentially a consequence of the fact that automatically
generated ids and user specified ids are different. Since the convertId()
method is used to both specify an id for a new vertex, and also to check that
the created vertex really does have the Id that's specified via comparison with
the return value of Vertex.id(), the convertId() method must be aware of
RAMCloudGraph's underlying implementation of user supplied Ids.