Fixed properties serialization bug.
A prior commit fixed a property serialization bug whereby strings with special
characters would not get their total length in bytes calculated correctly. This
fix, however, was only made in one place (the calculation of total bytes needed
to store the serialized version of the properties), but not in the place that
fills the length fields of this buffer with the correct value of the length of
the string that follows in the buffer. This commit fixes that problem so that
deserialization works correctly.