Changed ID generation scheme.
ID space is divided evenly into two 127 bit ID spaces. The upper range is for
auto generated IDs, whereas the lower range (with most significant bit 0) is
for user supplied IDs. Auto generated IDs in the upper range are generated by
randomly selecting 1 of NUM_ID_COUNTERS counters stored in RAMCloud and
atomically incrememting it. The new count value becomes the lower 64bit portion
of the new ID, and the upper 64bits is the index of the counter that was
incrememted. And again, for these auto generated IDs the 128th bit is set to 1.

This change is motivated by the fact that the LDBC benchmark uses unique 64bit
IDs *per* entity type. This means that we need 64bit ID spaces for each type of
entity in the graph. 128 bits ought to give us enough room ;)