Added UInt128 as TorcVertex internal ID rep.
Created a new class, UInt128, which stores an unsigned 128 bit integer. This
class is used as the internal ID representation for Torc vertices. It provides
several constructors to create a UInt128 from various other data types,
including Byte, Short, Integer, Long, String, BigInteger, and byte[]. Also
provides a toByteArray() method, the output of which is used in constructing
RAMCloud keys for various vertex objects, and can be fed back to the
constructor to create an exact copy of the UInt128.