Created new TorcEdge ID representation.
Created inner class in TorcEdge that represents the edge ID. An instance of
this inner class is returned on a call to TorcEdge.Id(). This object overrides
the basic methods of the Object class, including toString() and equals(), in
order to see and compare two TorcEdge IDs.

Also added a few other things:
 - Added new methods in TorcEdge for getting the vertex IDs and the type of the
edge.
 - TorcGraph has a new equals() method that defines equality between two
TorcGraph instances to mean that they are pointing at the same database.
 - Decided to return undirected edges in the implementation of Graph.edges()
method when no edgeIDs are given to the method.
 - Added more comments.
 - For Edge.vertices(), throws a RuntimeException if the method is called with
a valid direction but an undirected edge, in which case the operation doesn't
make sense.
 - Removed many of the TorcHelper methods that pertained to the ad-hoc byte[]
representation of Edge IDs.