Disabled user supplied IDs.
After trying to implement this, decided not to support it. Users can still
supply IDs of various types to addVertex and vertices() methods, however
these IDs are internally converted to a UInt128, and such is returned upon any
call to Vertex.id(). To truly support string IDs and so on would require using
these types as the internal type, and unfortunately that's too much work for
now.