Fixed deleteDatabaseAndCloseConnection.
This is issue #1. This method was supposed to clean up all the outstanding
transaction state. This is important because RAMCloudTransactions represent a
C++ RAMCloud Transaction object over JNI, and it needs to be freed when the
Java side object is closed, otherwise there will be memory leaks.  However, it
was only cleaning up the transaction state of the current thread. To free the
state of all threads, a new data structure was needed to be able to iterate
through it and close all other thread's outstanding transactions (ThreadLocal
has no such methods).