Added TorcVertexEdgeList.
A TorcVertexEdgeList is a RAMCloud-resident data structure representing a list
of edges of a certain TorcVertex that have the same label and direction from
the vertex. It abstracts details of the RAMCloud storage layer, and provides a
set of high-level methods for adding, fetching, and removing edges from the
list. These operations are all performed within a transaction context, and
therefore writes to this data structure are only externally visible after a
successful commit of the transaction. The implementation has been performance
tuned to make optimal use of the RAMCloud storage layer.