Made number of masters config param optional.
Sometimes a TorcGraph instance is created against an existing graph. In
this case the number of master servers is not under our control, it is
in the control of the RAMCloud cluster. In this case this detail is
abstracted away from us, and we should not need to provide this
parameter. If this is creating a new graph, however, then this parameter
should be provided to specify the server span of the graph.