Added RAMCloud Image Creation Mode.
When in RAMCloud image creation mode, methods such as loadVertex and
loadEdgeList write out RAMCloud key value pairs to a binary file on local disk
for loading into RAMCloud at a future date. This is very useful in situations
where a dataset graph is loaded many times over into RAMCloud. If the graph is
large, loading it via the TorcDB API can be prohibitively slow.