Made writeListToFile more efficient ico no props.
Sometimes edge lists are very very large, and have no properties. In this case,
the property list was being filled with empty array lists, all of which were
allocated on the stack and essentially useless. Modified the writeListToFile
method to interpret an empty property list as signaling that these edges have no
properties.