Fixed bug in TorcEdgeList.batchRead.
Previous update to the RAMCloud bindings made it so that
RAMCloudTransactionReadOp.getValue() returns null if the object doesn't exist,
whereas the previous behavior was to throw an exception. Looks like a piece of
code in TorcEdgeList.batchRead did not get properly updated, and was relying on
catching an exception, but didn't and ended up using a null returned object.