Added LDBC SNB optimized edge traversal methods.
Via the standard TinkerPop API the client has no way of querying for vertices of
a particular type when asking for all the neighbors of a vertex with a given
edge type. Since TorcDB stores edge lists not only by edge type, but also vertex
type, TorcDB must make extra reads to service queries that do not specify the
vertex type.

I extended the API to include a new method that includes the neighbor label the
client is interested in. This reduced the number of reads for servicing
LdbcShortQuery5 from 2 to 1.