Changed TorcGraphPOS to search ahead for labels.
For outE() and inE() steps, their labels do not appear until down the traversal
after an outV() or inV() step. Added a limited depth search for hasLabel here.

This is really a total hack.... I'm isolating the hackiness to this custom
ProviderOptimizationStrategy. The Gremlin query itself is still stock Gremlin
and will work with the default strategies.