Fixed bug in TorcHelper.intersect.
TorcHelper.intersect was updating TraversalResult.vMap, but not updating
TraversalResult.vList. Changed TraversalResult.vList to a vSet instead, and
added the correct operation.

The reason for changing to a Set is that this is data structure makes operations
more efficient in general, because this is used for getting the unique set of
neighbors touched by this traversal. Later, when we need to use that set to
intersect (or later, filter), it's more efficient to use a set than use a list.