Simplified traversal API to just "traverse()".
Instead of having getVertices and getEdges, now we have just the "traverse"
method which has an extra argument "fillEdge" which will fill in a data
structure in TraversalResult that describes the properties for all the edges.
This data structure is set to null with fillEdges = false.

Updated all the helper methods to work with this new framework.