Fixed bug with edge list keys.
I accidentally used serializeObjet() for serializing Strings used in edge list
keys (edge label and vertex label keys). For keys, the serialization format does
not need the overhead (i.e. type codes) used in real serialization when we need
to parse out the object and don't know what's in the bytes when we parse it. For
keys, the bytes are never deserialized anyways.