Optimization on single vertex property.
Map objects take up a LOT of memory when storing just a single property, as is
often the case in the LDBC social network benchmark. So in the case that we
really want to fetch only a single property, we can now specify that property to
fillProperties(), and then we will store this property in key/value pair member
variables natively in a TorcVertex object, instead of creating a Map<Object,
Object> properties object. Then when getting properties on the vertex, these
variables are checked.